Extract EnglishMnemonic into separate file
