+++ packages/iov-crypto/types/englishmnemonic.d.ts
@@ -1,0 +1,6 @@
+export declare class EnglishMnemonic {
+    private static readonly mnemonicMatcher;
+    private readonly data;
+    constructor(mnemonic: string);
+    asString(): string;
+}
